Fix warning: MobileCoreServices has been renamed.
Thanks to github/ElfSundae for opening the PR to fix this issue in 'master'. The PR was opened two months ago and has seen no reply from an AgileBits dev, which seems normal for them.
https://github.com/agilebits/onepassword-app-extension/pull/423

github/mccarron asked if this change could be in the 'add-framework-support' branch, also to no reply.
https://github.com/agilebits/onepassword-app-extension/pull/422

Given those delays, this commit mirrors ElfSundae's change, but on top of the 'add-framework-support' branch.